Switch webdriver to Firefox/Gecko and error catching
This switches the Selenium main webdriver to use Firefox/gecko as
the browser/engine for scraping purposes. The goal is for this to
run faster and enable the server to exit without leaving dangling
processes, which has been an issue with chrome-driver. Also added
are several error catching/timing fixes to improve performance.